Added safeguards before deletion

+++ main.py
@@ -4,6 +4,7 @@ import requests
 #Get users input on date range
 createdTo = input("Please input the first date you'd like to begin deleting links in YYYY-MM-DD format: ")
 createdFrom = input("Please input the last date you'd like to delete links in YYYY-MM-DD format: ")
+print("Retrieving links to delete. Please wait.")
 
 #API urls
 capsuleList = "https://www.capsulink.com/api/list"
@@ -24,6 +25,7 @@ linksResponse = requests.get(capsuleList, headers=headers, params=params)
 #Initialize dictionary for those links, and list to hold IDs
 dictOfCapsules = {}
 capsuleIDs = []
+linksToDelete = []
 
 #Convert links into json format in order to parse
 if linksResponse.status_code == 200:
@@ -33,16 +35,24 @@ if linksResponse.status_code == 200:
 for i in range(len(dictOfCapsules['data'])):
     capsuleIDs.append(dictOfCapsules['data'][i]['id'])
 
-print('Thank you! Links have been retrieved. Beginning deletion, please keep this window open.')
+#Extract links for final confirmation
+for i in range(len(dictOfCapsules['data'])):
+    linksToDelete.append(dictOfCapsules['data'][i]['redirect_url'])
 
-#Parse the link IDs and delete one by one    
-for linkID in capsuleIDs:
-    data = {
-        "id": linkID
-    }
+#List out links for user to review
+print(f'LINKS TO BE DELETED: {linksToDelete}')
+
+userFinalConfirm = input("Please confirm if you'd like these links deleted (y/n): ")
+
+#Parse the link IDs and delete one by one if user has authorized
+if userFinalConfirm == 'y':
+    for linkID in capsuleIDs:
+        data = {
+            "id": linkID
+        }
+        
+        deleteResponse = requests.delete(capsuleDeleter, headers=headers, json=data)
+        if deleteResponse.status_code != 200:
+            print(f'Error deleting link with ID number: {linkID}')
     
-    deleteResponse = requests.delete(capsuleDeleter, headers=headers, json=data)
-    if deleteResponse.status_code != 200:
-        print(f'Error deleting link with ID number: {linkID}')
-
-print(f'Links from {createdTo} to {createdFrom} have successfully been deleted.')
+    print(f'Links from {createdTo} to {createdFrom} have successfully been deleted.')
